Fix extract tasks to actually be properly lazy
